Trait buildouts for cipher building and param parsing.
