+++ src/Crypto/AbstractCryptoClient.php
@@ -32,13 +32,8 @@ abstract class AbstractCryptoClient
      *                           or decrypting.
      * @param int $keySize Size of the encryption key, in bits, that will be
      *                     used.
-     *
-     * @return string
      */
-    protected function getCipherOpenSslName($cipherName, $keySize)
-    {
-        return "aes-{$keySize}-{$cipherName}";
-    }
+    abstract protected function getCipherOpenSslName($cipherName, $keySize);
 
     /**
      * Constructs a CipherMethod for the given name, initialized with the other
@@ -49,22 +44,9 @@ abstract class AbstractCryptoClient
      * @param int $keySize Size of the encryption key, in bits, that will be
      *                     used.
      *
-     * @return CipherMethod
-     *
      * @internal
      */
-    protected function buildCipherMethod($cipherName, $iv, $keySize)
-    {
-        switch ($cipherName) {
-            case 'cbc':
-                return new Cbc(
-                    $iv,
-                    $keySize
-                );
-            default:
-                return null;
-        }
-    }
+    abstract protected function buildCipherMethod($cipherName, $iv, $keySize);
 
     /**
      * Performs a reverse lookup to get the openssl_* cipher name from the
@@ -72,22 +54,9 @@ abstract class AbstractCryptoClient
      *
      * @param $aesName
      *
-     * @return string
-     *
      * @internal
      */
-    protected function getCipherFromAesName($aesName)
-    {
-        switch ($aesName) {
-            case 'AES/GCM/NoPadding':
-                return 'gcm';
-            case 'AES/CBC/PKCS5Padding':
-                return 'cbc';
-            default:
-                throw new \RuntimeException('Unrecognized or unsupported'
-                    . ' AESName for reverse lookup.');
-        }
-    }
+    abstract protected function getCipherFromAesName($aesName);
 
     /**
      * Dependency to provide an interface for building an encryption stream for
